Add explicit typecasts to ui/accessibility enums.
This patch is in preparation for switching the accessibility
enums in ui/accessibility/ax_enums.idl to a .mojom file instead.
Because the generated C++ headers from a .mojom file use enum
class instead of enum, we need to explicitly cast anytime we
use an enum as an int.

Bug: 650275

+++ chrome/browser/chromeos/arc/accessibility/ax_tree_source_arc.cc
@@ -270,12 +270,14 @@ void PopulateAXState(arc::mojom::AccessibilityNodeInfoData* node,
         GetBooleanProperty(node, AXBooleanProperty::CHECKED);
     const ui::AXCheckedState checked_state =
         is_checked ? ui::AX_CHECKED_STATE_TRUE : ui::AX_CHECKED_STATE_FALSE;
-    out_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE, checked_state);
+    out_data->AddIntAttribute(ui::AX_ATTR_CHECKED_STATE,
+                              static_cast<int32_t>(checked_state));
   }
 
   if (!GetBooleanProperty(node, AXBooleanProperty::ENABLED)) {
-    out_data->AddIntAttribute(ui::AX_ATTR_RESTRICTION,
-                              ui::AX_RESTRICTION_DISABLED);
+    out_data->AddIntAttribute(
+        ui::AX_ATTR_RESTRICTION,
+        static_cast<int32_t>(ui::AX_RESTRICTION_DISABLED));
   }
 
   if (!GetBooleanProperty(node, AXBooleanProperty::VISIBLE_TO_USER)) {
