+++ content/browser/accessibility/browser_accessibility_manager_unittest.cc
@@ -598,7 +598,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRange) {
   inline_text1.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text1.location = gfx::RectF(100, 100, 29, 9);
   inline_text1.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets1;
   character_offsets1.push_back(6);   // 0
   character_offsets1.push_back(11);  // 1
@@ -617,7 +617,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRange) {
   inline_text2.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text2.location = gfx::RectF(100, 109, 28, 9);
   inline_text2.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets2;
   character_offsets2.push_back(5);
   character_offsets2.push_back(10);
@@ -684,7 +684,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeMultiElement) {
   inline_text1.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text1.location = gfx::RectF(0, 20, 33, 9);
   inline_text1.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets { 10, 21, 33 };
   inline_text1.AddIntListAttribute(
       ui::AX_ATTR_CHARACTER_OFFSETS, character_offsets);
@@ -703,7 +703,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeMultiElement) {
   inline_text2.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text2.location = gfx::RectF(10, 40, 33, 9);
   inline_text2.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   inline_text2.AddIntListAttribute(
       ui::AX_ATTR_CHARACTER_OFFSETS, character_offsets);
   static_text2.child_ids.push_back(5);
@@ -793,7 +793,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeBiDi) {
   inline_text1.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text1.location = gfx::RectF(100, 100, 30, 20);
   inline_text1.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets1;
   character_offsets1.push_back(10);  // 0
   character_offsets1.push_back(20);  // 1
@@ -808,7 +808,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeBiDi) {
   inline_text2.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text2.location = gfx::RectF(130, 100, 30, 20);
   inline_text2.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_RTL);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_RTL));
   std::vector<int32_t> character_offsets2;
   character_offsets2.push_back(10);
   character_offsets2.push_back(20);
@@ -870,7 +870,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeScrolledWindow) {
   inline_text.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text.location = gfx::RectF(100, 100, 16, 9);
   inline_text.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                              ui::AX_TEXT_DIRECTION_LTR);
+                              static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets1;
   character_offsets1.push_back(6);   // 0
   character_offsets1.push_back(11);  // 1
@@ -940,7 +940,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeOnParentElement) {
   inline_text1.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text1.location = gfx::RectF(100, 100, 40, 20);
   inline_text1.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets1;
   character_offsets1.push_back(20);  // 0
   character_offsets1.push_back(40);  // 1
@@ -953,7 +953,7 @@ TEST(BrowserAccessibilityManagerTest, BoundsForRangeOnParentElement) {
   inline_text2.role = ui::AX_ROLE_INLINE_TEXT_BOX;
   inline_text2.location = gfx::RectF(160, 100, 40, 20);
   inline_text2.AddIntAttribute(ui::AX_ATTR_TEXT_DIRECTION,
-                               ui::AX_TEXT_DIRECTION_LTR);
+                               static_cast<int32_t>(ui::AX_TEXT_DIRECTION_LTR));
   std::vector<int32_t> character_offsets2;
   character_offsets2.push_back(20);  // 0
   character_offsets2.push_back(40);  // 1
